clean up and rename
